Enhancement/#3 linking second device (#10)
Linking between Signal-Desktop devices was added.
There are two new major views, the manage_devices_view and the add_device_view.
In the former one devices can be removed and in the second view new ones can be added.
To deal with the device management both the AccountManager and the WebAPI had to be extended.

Resolves #3