Enhancement/#5 name change (#11)
With this commit, it is possible to change a contact's name.
For this a menu option "Edit Name" is added to the conversation menu in the upper right corner.
When clicked, the name changes to an input field and the name can be changed.
For the input field a new Component NameInput is introduced.
The conversation model is also expanded by a method for setting it's name.

Resolves #5